fix: script refactoring
